perf: reduce server memory by lazy-loading VAD and using half-precision model
- Make Silero VAD loading lazy instead of unconditional at startup
- Load ResNet34 model in float16 (half-precision) to halve parameter memory
- Auto-cast fbank features to match model dtype for safe inference
- Return embeddings as float32 for downstream compatibility

+++ src/wespeaker_deep_edge/wespeaker_deep_dege.py
@@ -106,7 +106,7 @@ class WespeakerDeep:
 
 
         model_dir = Path(__file__).parent / "_models" / "vblinkf"
-        self._model = wespeaker.load_model(str(model_dir))
+        self._model = wespeaker.load_model(str(model_dir), dtype="float16")
         self._model_dir = str(model_dir.resolve())
         self.sample_rate = sample_rate
         self._deep_config = config if config is not None else DeepConfig()
